+++ src/templates/index.ts
@@ -1,1 +1,0 @@
-export * from "./stacked_view_view_model";