+++ src/templates/stacked_view_view_model/index.ts
@@ -1,2 +1,0 @@
-export * from "./stacked_view.template";
-export * from "./stacked_view_model.template";